junction-core: add error types for xDS
